Adds cypherLoading state to MotifPanel.
This is used to disable the submit button while the cypher is being
loaded from the server. After the processRequest function is called with
the cyper payload, the isQuerying prop can be used to disable the submit
button. Without the addition of this to the state, the submit button is
active and can be clicked until the cypher has been retrieved.